isolate isce3 in testing
